Get Simulation Parameters

Get Simulation Parameters

Usage

ExtractParameters(sim_object, steps = NULL)

Arguments

sim_object

simulation object created with CreateSimulationObject

steps

which parameters to extract from the SpatSimObj

Value

a list of S3 clas SimParams containing parameters stored in the sim_object

Details

This function will return any paramters that are stored in the simulation object. If no simulation steps have been run, then this will return the default parameters. The defaults are over written if new parameters are provided at each step.

Examples

#create simulation object
sim_obj = CreateSimulationObject()
#> No `window` specified - defaulting to x (0, 10); y (0, 10)
#extract default paramters for the Tissue simulation step
defs = ExtractParameters(sim_obj, "Tissue")
#> 	Tissue has not yet been simulated